+++ programs/uxd/src/instructions/credix_lp/mod.rs
@@ -1,15 +1,13 @@
 pub mod collect_profits_of_credix_lp_depository;
 pub mod edit_credix_lp_depository;
 pub mod mint_with_credix_lp_depository;
-pub mod rebalance_redeem_from_credix_lp_depository;
-pub mod rebalance_request_from_credix_lp_depository;
+pub mod rebalance_from_credix_lp_depository;
 pub mod redeem_from_credix_lp_depository;
 pub mod register_credix_lp_depository;
 
 pub use collect_profits_of_credix_lp_depository::*;
 pub use edit_credix_lp_depository::*;
 pub use mint_with_credix_lp_depository::*;
-pub use rebalance_redeem_from_credix_lp_depository::*;
-pub use rebalance_request_from_credix_lp_depository::*;
+pub use rebalance_from_credix_lp_depository::*;
 pub use redeem_from_credix_lp_depository::*;
 pub use register_credix_lp_depository::*;
