--- programs/uxd/src/utils/credix_lp/mod.rs
@@ -1,5 +1,0 @@
-pub mod calculate_profits_value;
-pub mod calculate_target_overflow_value;
-
-pub use calculate_profits_value::*;
-pub use calculate_target_overflow_value::*;
