+++ programs/uxd/src/utils/mod.rs
@@ -1,12 +1,10 @@
 pub mod calculate_amount_less_fees;
-pub mod credix_lp;
 pub mod math;
 pub mod validate_collateral_amount;
 pub mod validate_collateral_mint_usdc;
 pub mod validate_redeemable_amount;
 
 pub use calculate_amount_less_fees::*;
-pub use credix_lp::*;
 pub use math::*;
 pub use validate_collateral_amount::*;
 pub use validate_collateral_mint_usdc::*;
