+++ programs/uxd/tests/integration_tests/api/program_uxd/instructions/mod.rs
@@ -8,6 +8,7 @@ pub mod process_initialize_identity_depository;
 pub mod process_mint_with_credix_lp_depository;
 pub mod process_mint_with_identity_depository;
 pub mod process_mint_with_mercurial_vault_depository;
+pub mod process_rebalance_from_credix_lp_depository;
 pub mod process_redeem_from_identity_depository;
 pub mod process_redeem_from_mercurial_vault_depository;
 pub mod process_register_credix_lp_depository;
@@ -23,6 +24,7 @@ pub use process_initialize_identity_depository::*;
 pub use process_mint_with_credix_lp_depository::*;
 pub use process_mint_with_identity_depository::*;
 pub use process_mint_with_mercurial_vault_depository::*;
+pub use process_rebalance_from_credix_lp_depository::*;
 pub use process_redeem_from_identity_depository::*;
 pub use process_redeem_from_mercurial_vault_depository::*;
 pub use process_register_credix_lp_depository::*;
